wip(JsonForm): Restructured internals
